binary and linear search work. comparisons count properly. added comments and proper line spacing. need to round to 2 decimals

+++ C++/Assignment2/Searching.hpp
@@ -10,8 +10,8 @@ using namespace std;
 class Searching{
     public:
         //declaration of linear and binary search
-        int linearSearch(vector<string>& magicItems, vector<string>& randomItems);
-        int binarySearch(vector<string>& magicItems, vector<string>& randomItems);
+        float linearSearch(vector<string>& magicItems, vector<string>& randomItems);
+        float binarySearch(vector<string>& magicItems, vector<string>& randomItems);
 
 };
